Refactor types, helpers and hooks from ViewportSliceProgressScrollbar into a subdirectory of files. Update the loading className of the viewed fill to match the original design.
