+++ extensions/cornerstone/src/Viewport/Overlays/ViewportSliceProgressScrollbar/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ViewportSliceProgressScrollbar';
